+++ CHANGES.md
@@ -4,7 +4,9 @@
 2.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Bug 65: don't allow passing undefined namevars or undefined attributes
+  to the component constructor. Also prohibit (accidentally) overriding
+  methods.
 
 
 2.0 (2020-07-15)